mystifly_api/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
applicationinsights==0.11.10
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
awscli==1.25.59
azure-common==1.1.28
azure-core==1.25.1
azure-nspkg==3.0.2
azure-storage-blob==12.13.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.24.58
botocore==1.27.58
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.5
django-cronjobs==0.2.3
django-crontab==0.7.1
django-elasticsearch-dsl
django-extensions==3.2.0
django-log-viewer==1.1.5
django-prometheus==2.2.0
djangorestframework==3.13.1
docutils==0.16
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
frozenlist==1.3.3
gevent==22.10.2
google-auth==2.9.0
google-auth-oauthlib==0.5.2
greenlet==2.0.1
gspread==5.4.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
isodate==0.6.1
jmespath==1.0.1
JSON-log-formatter==0.5.1
Jinja2==3.1.2
kombu==5.2.4
msrest==0.7.1
multidict==6.0.2
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.1.0
opencensus==0.11.2
opencensus-ext-azure==1.1.9
opencensus-ext-django==0.8.0
opencensus-ext-postgresql==0.1.3
opencensus-ext-requests==0.8.0
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-logstash==0.4.8
pytz==2022.1
PyYAML==5.4.1
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.7
s3transfer==0.6.0
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.8.1
wrapt==1.14.1
wsproto==1.1.0
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.2
selenium-wire